Refactor SDK to make it modular
