+++ lib/client/getUrl.js
@@ -1,0 +1,9 @@
+const Client = require('../client/Client');
+/**
+ * Get API Base URL with API Version
+ * @private
+ * @returns {string} Base URL to Veryfi API
+ */
+Client.prototype._get_url = function () {
+    return this.base_url + "api/" + this.api_version;
+}
