+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@veryfi/veryfi-sdk",
-  "version": "1.3.2",
+  "version": "1.4.0",
   "description": "Node.js module for communicating with the Veryfi OCR API",
-  "main": "lib/main.js",
+  "main": "lib/client/client.js",
   "typings": "lib/types/main.d.ts",
   "scripts": {
     "test": "jest",
